+++ src/xyz/oihana/schema/constants/traits/AuthTrait.php
@@ -7,6 +7,7 @@ use xyz\oihana\schema\constants\traits\auth\InvitationTrait;
 use xyz\oihana\schema\constants\traits\auth\KeyfileTrait;
 use xyz\oihana\schema\constants\traits\auth\OAuthClientTrait;
 use xyz\oihana\schema\constants\traits\auth\PasswordResetTrait;
+use xyz\oihana\schema\constants\traits\auth\PendingRevocationTrait;
 use xyz\oihana\schema\constants\traits\auth\PermissionTrait;
 use xyz\oihana\schema\constants\traits\auth\PolicyTrait;
 use xyz\oihana\schema\constants\traits\auth\RoleTrait;
@@ -23,6 +24,7 @@ trait AuthTrait
         InvitationTrait ,
         OAuthClientTrait ,
         PasswordResetTrait ,
+        PendingRevocationTrait ,
         PermissionTrait ,
         PolicyTrait ,
         RoleTrait ,